FEATURE: add aws_dns attribute for ipv4addr
AmazonProvidedDNS is an Amazon DNS server, and this option enables DNS
for instances that need to communicate over the VPC's Internet gateway.
The string AmazonProvidedDNS maps to a DNS server running on a reserved
IP address at the base of the VPC IPv4 network range, plus two. For
example, the DNS Server on a 10.0.0.0/16 network is located at 10.0.0.2.